# Use encoding/binary to convert hashing seed for better performance