--- a/processor/probabilisticsamplerprocessor/fnvhasher_test.go
+++ b/processor/probabilisticsamplerprocessor/fnvhasher_test.go
@@ -1,0 +1,13 @@
+package probabilisticsamplerprocessor
+
+import (
+	"testing"
+)
+
+func BenchmarkSeedConversion(b *testing.B) {
+	val := uint32(0x3024001) // Just a random 32 bit int
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		i32tob(val)
+	}
+}