--- a/processor/probabilisticsamplerprocessor/fnvhasher_test.go
+++ b/processor/probabilisticsamplerprocessor/fnvhasher_test.go
@@ -1,3 +1,6 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
 package probabilisticsamplerprocessor
 
 import (